refactor: move isChained to utils
